Separate flight behavior from CesiumCameraController
